# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Helper to add price.js",
        "title": "GEO Support API",
        "contact": {},
        "version": "0.01"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/domains/{id}": {
            "post": {
                "description": "set_flag is out parameters: 0 - new(set automatic), 1 - active, 2 - ignore",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Set flag to domain",
                "operationId": "set-flag-domain",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "id domain",
                        "name": "id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "description": "set flag",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/modelsstruct.Flag"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "modelsstruct.Flag": {
            "type": "object",
            "required": [
                "set_flag"
            ],
            "properties": {
                "set_flag": {
                    "type": "integer"
                }
            }
        },
...
    }
}